multiverse runner: exact directory name matching
the current consensus is that when a suite name "filter" is passed to
the multiverse rake task, that filter should require an exact match with
a suite directory name.

for example, with the following command:

```shell
bundle exec rake test:multiverse[active_record]
```

the desired result is to have the `active_record` suite ran, but the
`active_record_pg` suite be excluded.