tests: allow call counts to be operator based
minitest: When defining an expected "call count" value, support the
inclusion of any one of the ">", "<", ">=", and "<=" operators instead
of always requiring an exact "==" style match.

to demonstrate this new functionality, the ActiveRecord test for
Postgres SELECT statements has been updated to use ">=" and permit the
extra schema/type cache query calls to be present and permitted by the
test.