rework cross agent pull/push rake tasks
The previous passing of `-u` to `rsync` (don't overwrite the destination
content if it has a newer timestamp than the source content) was
problematic because it assumed that a developer would have both the Ruby
agent and the cross agent specs cloned, then have new specs authored and
pulled down, and then perform the sync. If the cross agent specs haven't
been updated in a while, then the clone order of the repos becomes a
factor in the timestamp comparison.

Now, the pull and push rake targets will always overwrite the
destination content with the source content. To compensate for this, the
user is now prompted to continue or cancel with a warning that any
unsaved changes will be overwritten, and the previous source code
comment based warning is now presented to stdout as well.

resolves #924