update Puma multiverse tests
update the testing of our Puma instrumentation:

- point to the latest package of each of the 3 most recently released
  major Puma versions
- stop having newer Rubies test multiple versions of Rack. Now simply
  test with the latest Rack for those Rubies that can (Ruby 2.3+) and
  the latest Ruby 2.2 compatible Rack for Ruby 2.2
- add a test for Puma URLMap handling which was previously instrumented
  but not covered by multiverse testing

Pumas are less than half as fast as cheetahs and over 1.5 times as
heavy.

But Pumas can live 19 years as opposed to 11 for cheetahs. If the Ruby
Puma project had been named Cheetah when it began in 2011, it might have
been on its last legs now in 2022.

resolves #1189