Update Envfile format
The Envfiles now consist of:
1. An array of versions of the subject of the suite that will be tested.
This often includes in index 1 and index 2 the lowest and highest
compatible versions of Ruby

2. A call to a method, `#gem_list`, with an argument of the gem version
and returns a heredoc with the gems that should be applied for the suite

3. A call to create_gemfiles with the version array and the gem_list
method passed as arguments

Envfiles for gems whose instrumentation will be deprecated or undergo
version deprecation in the next major release were not touched.

The SimpleCov threshold was also lowered to account for drops in
coverage percentages with this change.