Execute @@after_run check only for Ruby >= 2.7
This block was being run on versions of Ruby that do not generate
SimpleCov coverage results, causing an error to be raised
on the CI run.

This if statement should prevent the errors from being raised.

Also, the rescue has been limited to NameError instead of globally
rescuing any exception.