Update gRPC instrumentation supportability_name
Previously, when gRPC instrumentation was installed, the log message
not include "gRPC". This is because the
name used in the metrics and logs is by default derived from the second
to last module namespace. In this case, it was coming up as "Client" or
"Server".
By passing a third, optional argument to
DependencyDetection::Dependent#prepend_instrument and #chain_instrument
we are able to get more precise log messages and Supportability metrics.