gRPC client: don't touch initialize
for the gRPC client instrumentation, we had been prepending/chaining
`initialize` calls to get at the host argument in order to apply our
gRPC host denylist behavior. when other libraries are themselves
bringing in the gRPC client library (via the `grpc` gem) and using their
own `initialize` chaining and `super` calls, we run the risk of passing
the wrong number of arguments to the `initialize` method in the chain
that we are handing things off to.

notably, this causes `google-ads-googleads` based apps to encounter
issues.

to address this, let's get out of the business of touching `initialize`
entirely, and source the host information from the gRPC client
instance's own `@host` ivar.

resolves #1337