Add elasticsearch instrumentation (#1525)
Introduces instrumentation for the elasticsearch gem versions 7.x and 8.x and add an elasticsearch multiverse suite to the database multiverse group. 

The Mongo nosql obfuscator has been moved out of mongo and is now a general NoSql obfuscator used by both mongo and elasticsearch. 

Instrumentation generator thor task is updated with fixes that were found when using the generator to create the elasticsearch files. 

Elasticsearch 7 and 8 are now included in both the ci.yml and ci_cron.yml workflows to support the elasticsearch multiverse suite running on the CI. The docker compose file has also been updated to run elasticsearch 7 on port 9200 and elasticsearch 8 on port 9250 for local testing.