Correctly identify Unicorn when used with Rack v3
With version 3.0.0 of Rack in play, the agent can be fooled into
thinking that web dispatching duties are being handled by Webrick. This
is because the agent performs a less strict test for Webrick compared to
other dispatchers; checking only for the existence of related namespaces
to be defined.

With these changes, the agent will now treat Webrick as the last
dispatcher checked after all other anticipated production use
dispatchers have not been located. Specifically, this means that
Webrick is checked after Unicorn, Rainbows, and FastCGI; all of which
may have been incorrectly reporting as having Webrick as the dispatcher
if Rack v3.0.0 was being used.

Resolves #1394