disable, make private ActionDispatch notifications
the ActiveSupport event notifications based instrumentation for
ActionDispatch overlaps too greatly with our existing Rails middleware
instrumentation, leading to confusing and potentially costly overlap for
Rails users.

ideally, we'd like to standardize on leveraging Rails notifications for
all Rails instrumentation and remove any custom Rails instrumentation
that overlaps. But in this case with ActionDispatch the notifications
aren't available until Rails v6, and we aren't prepared at this time to
stop supporting earlier Rails versions.

so for now, retain the notifications based instrumentation for
ActionDispatch but have it configured - privately - to be disabled by
default.