MaxPayloadSizeLimit supportability metric rename
As per internal agent specs PR 585, make sure the `MaxPayloadSizeLimit`
supportability metric includes the agent language name ("Ruby").

NOTE that the `MaxPayloadSizeLimit` part of the name itself isn't great,
but we've decided to keep it as-is for now. The metric actually keeps a
running tally of violations of the size limit and doesn't actually
report what the configured limit is.