Efficiency improvements for segment completion
When lots of instrumented activity takes place nested within an outer
instrumented action, the number of child segments belonging to a parent
segment may reach a quantity (likely 10,000 or higher) that causes the
agent to consume considerable CPU resources and either take a long time
or hang indefinitely while processing the child segments at parent
segment finish time.

Examples of nested activity are:

- Instrumented data store operations that occur within an instrumented
  background job process
- Bulk data load operations that take place during a public API method
  call to trace the execution of a block of code

Version 9.0 of the agent began enabling the
`instrumentation.thread.tracing` config parameter by default. The
agent's `Thread` based tracing controlled by this parameter allows the
agent to record metrics for activity taking place within Ruby threads
that would otherwise be hidden from view of the agent. While this
additional visibility is generally a great enhancement of the agent's
detailed observability capabilities, it can mean that nested child
segments are created by threads spawned by long running processes that
are themselves instrumented by the agent.

After studying exactly what the agent does with child segments in
quantities of 20,000 to 100,000, the agent's maintainers noticed that
the parent segment spends a lot of time looking for start/finish time
overlaps between all child segments and itself. This is critically
important for the creation of accurate timing information when child
segments operate concurrently, so as to not double report the time spent
in the overlap periods. But when large quantities of child segments are
present to be processed - even if they don't feature any overlap at all
in timing - the inefficiencies in the overlap calculations can be
lengthy and resource intensive.

With the changes included here, the process of calculating child segment
overlap is now thought to be dramatically more efficient without
sacrificing any accuracy.