Active Record notifications: permit disabling
To prevent the agent from subscribing to Active Record notifications and
otherwise instrumenting Active Record, the configuration parameter
`:disable_active_record_instrumentation` should be set to a value of
`true` in the `newrelic.yml` config file.

This change permits that parameter to function as expected by fixing a
typo in which the code expected `:disable_activerecord_instrumentation`
without an underscore. Note that prior to this change, putting
`:disable_activerecord_instrumentation` in the config file would also
not work, because the config parameter is spelled correctly in
`default_source.rb`. This change is needed for the parameter to function
at all.