Add log level filtering
This creates a new configuration, application_logging.
forwarding.log_level. This configuration defines the
minimum log level a log event should have to be forwarded to
New Relic.

Filtering is done in the LogEventAggregator#record method.

The default is "debug" which will forward all logs.

Setting the configuration to another built-in Ruby log level, such as
"error", will forward only logs with a severity of error or
above. This would mean logs with "error", "fatal" or "unknown"
would be forwarded to New Relic.

If an invalid log level is given to the configuration, the agent
will default to forwarding all logs/having a "debug" level.

Log levels outside the standard Ruby Logger::Severity constants
will always have their events forwarded.