Create config_docs workflow
This workflow leverages the newrelic:config:docs[html] rake command to
generate the Ruby agent configuration document for the New Relic docs
website with values from the current default source.

After the config content file is generated, a PR is created on the
newrelic/docs-website with the changes using the newrelic-ruby-agent
-bot.

It is intended to run on release. Workflow dispatch is another
option to allow docs contributors to easily update the page content.