Update elasticsearch port_path_or_id value
Previously, the `port_path_or_id` value for Elasticsearch segments
was set to the path given to the `request_with_tracing` method. This
caused a metrics grouping issue for a customer because a new
instance metric ("Datastore/instance/Elasticsearch/<host>/<port>")
was created for every Elasticsearch document ID.

The source of the host value is a hash that also contains a port key. This
commit updates the value of `port_path_or_id` to use the port from the
`nr_hosts` hash.