Ruby 3.3+ fix for .gemspec test
Ruby 3.3.0-preview2 causes an '(eval ...' String to be yielded for a
__FILE__ call performed within an `eval`. This String won't work
properly in contexts that expect a String representing a valid
filesystem path.

We don't want to touch the .gemspec file for the sake of the test, and
we don't yet want to refactor out the `eval` call that the test uses, so
for now let's just make sure the test never has to deal with a __FILE__
call within the `eval`.