Support cgroups v1 and v2 for Docker containers
Previously the agent was only capable of gleaning a Docker container id
by leveraging a cgroups v1 approach. Now, it will first attempt to
leverage a cgroups v2 approach and fall back to the v1 approach if that
doesn't work. All other Docker related behavior such as that seen with
containers that don't have cgroups functionality enabled is not
impacted.