CI: fix flapping Sidekiq test
The singleton instance of `Sidekiq::CLI` needs to have a Sidekiq
configuration set on it for certain operations to work.

Previously we had been lucking out by having the `cli` test helper
method be called prior to the `test_captures_sidekiq_internal_errors`
test (which does not make use of the helper) being ran.

To prevent flapping, have the test make use of the `cli` helper (which
preps a CLI instance with a config) to make it always pass regardless of
the run order.