Ethon: focus on return_code for errors
For an instance of `Ethon::Easy`, consider a `#return_code` result not
equal to `:ok` to be the indicator of an error instead of the
`#response_code` being equal to `0`. With direct Ethon usage, these
equality checks seem interchangeable but with Typhoeus the underyling
easy object is left with a 0 for its response code on success.