ActiveSupport notifications: allowlist event names
Added a new :'instrumentation.active_support_notifications.active_support_events'
configuration parameter to permit a user defined allowlist of Active
Support notifications event names for the agent to subscribe to that are
related to the Active Support library itself.

Event names that are not included on the list will be ignored by the
agent.