Remove rake and runner from denylist default
The inclusion of Rails::Command::RunnerCommand and
Rails::CommandRakeCommand prevented the agent from starting for
background jobs in an application with ActiveJob and Solid Queue.

We encourage users who do not want the agent to run with `rails runner`,
`rails db:*`, or other rake tasks defined in Rails to include these
constants in their `autostart.denylisted_constants` configuration.