JRuby fix for concurrency issue with manager cache
The configuration manager's alteration of its internal cache hash to
preserve previously dynamically defined values has been known to cause
problems with JRuby given its concurrent access to the hash object. We
have used `synchronize` and `Hash#dup` fixes previously, but still have
reports of issues.

When it comes down to it, the alteration of the hash was only added and
still only needed for use by the security agent. So let's just not even
attempt to alter the hash if JRuby is in play but the security agent is
not.